Localize weather condition strings
Since we use the free version of the API and therefore want to make
sure we stay under the rate limit as much as possible, we take the
burden of localizing the conditions upon ourselves. We could fetch it
via the API, but that ruins our cache strategy if we don't have your
locale cached.

I used the data from here https://openweathermap.org/weather-conditions
to develop our own set of strings, which we will wait to have translated